requirements.txt
alembic==0.9.9
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
psycopg2-binary==2.7.4
six==1.11.0
sqlalchemy==1.2.7
testing.common.database==2.0.3
testing.postgresql==1.3.0
Werkzeug==0.12.2
